Fix spelling error (pimitive -> primitive)
